Move computation part to presenter.
