+++ src/main/java/oripa/gui/view/estimation/EstimationResultUIView.java
@@ -18,6 +18,10 @@
  */
 package oripa.gui.view.estimation;
 
+import java.util.Map;
+import java.util.Set;
+import java.util.function.Function;
+
 import oripa.domain.fold.FoldedModel;
 import oripa.domain.fold.origeom.OverlapRelation;
 import oripa.gui.view.View;
@@ -55,6 +59,13 @@ public interface EstimationResultUIView extends View {
 	void addExportButtonListener(Runnable listener);
 
 	void addSaveSVGCofigButtonListener(Runnable listener);
+
+	/**
+	 * @param listener
+	 *            Output type: < index of subface, map< index of order, overlap
+	 *            relation indices > >
+	 */
+	void setFilterInitializationListener(Function<FoldedModel, Map<Integer, Map<Integer, Set<Integer>>>> listener);
 
 	void showExportErrorMessage(Exception e);
